rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "datasets>=4.8.4",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain-ollama>=1.1.0",
    "ollama>=0.6.1",
    "onnxruntime",
    "openai>=2.31.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.2.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "ragas>=0.4.3",
    "sqlalchemy>=2.0.49",
    "tabulate>=0.10.0",
    "unstructured[all-docs]>=0.18.32",
    "uvicorn>=0.44.0",
]

[project.scripts]
rag = "rag.cli:main"

[dependency-groups]
dev = [
    "black>=22.3.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/rag"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
